+++ qwen3/__init__.py
@@ -1,0 +1,1 @@
+"""Qwen3 package"""
